moved all of the mocked responses into one file
