--- cypress/e2e/communities/Like-response.ts
@@ -1,7 +1,0 @@
-export const LIKE_RESPONSE = {
-  id: 'bf59fb3f-c07c-465d-9303-33a15ab507ae',
-  identity_id: '20b07c1c-b142-4ee3-a87f-122aa99a7659',
-  post_id: 'bf59fb3f-c07c-465d-9303-33a15ab507ae',
-  created_at: '2025-03-11T12:30:43.974Z',
-  comment_id: null,
-};
