--- cypress/e2e/communities/notificationMock.ts
@@ -1,998 +1,0 @@
-export const Notification = {
-  page: 1,
-  limit: 50,
-  total_count: 12,
-  items: [
-    {
-      id: '723022ee-83ed-4f7c-b119-13f006d4cbd9',
-      type: 'CHAT',
-      ref_id: 'eb19f03b-b08a-4177-a9d8-2d8059899060',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        id: 'eb19f03b-b08a-4177-a9d8-2d8059899060',
-        body: {
-          body: 'OrganizationTest sent you a new message',
-          title: 'New Message',
-        },
-        text: 'test test',
-        type: 'CHAT',
-        media: null,
-        muted: false,
-        orgin: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          created_at: '2025-02-27T10:59:24.278Z',
-        },
-        refId: 'eb19f03b-b08a-4177-a9d8-2d8059899060',
-        chat_id: '47909b77-87b4-461d-b242-eb63bd498fab',
-        replied: false,
-        identity: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T13:16:43.884Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: '47909b77-87b4-461d-b242-eb63bd498fab',
-        reply_id: null,
-        created_at: '2025-03-05T16:45:26.175Z',
-        deleted_at: null,
-        updated_at: '2025-03-05T16:45:26.175Z',
-        identity_id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-        consolidate_number: 0,
-      },
-      view_at: '2025-03-05T16:45:33.770Z',
-      read_at: '2025-03-05T16:45:40.341Z',
-      updated_at: '2025-03-05T16:45:40.341Z',
-      created_at: '2025-03-05T16:45:26.228Z',
-      silent: false,
-    },
-  ],
-};
-
-export const NOTIFICATIONS = {
-  page: 1,
-  limit: 50,
-  total_count: 12,
-  items: [
-    {
-      id: '723022ee-83ed-4f7c-b119-13f006d4cbd9',
-      type: 'CHAT',
-      ref_id: 'eb19f03b-b08a-4177-a9d8-2d8059899060',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        id: 'eb19f03b-b08a-4177-a9d8-2d8059899060',
-        body: {
-          body: 'OrganizationTest sent you a new message',
-          title: 'New Message',
-        },
-        text: 'test test',
-        type: 'CHAT',
-        media: null,
-        muted: false,
-        orgin: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          created_at: '2025-02-27T10:59:24.278Z',
-        },
-        refId: 'eb19f03b-b08a-4177-a9d8-2d8059899060',
-        chat_id: '47909b77-87b4-461d-b242-eb63bd498fab',
-        replied: false,
-        identity: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T13:16:43.884Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: '47909b77-87b4-461d-b242-eb63bd498fab',
-        reply_id: null,
-        created_at: '2025-03-05T16:45:26.175Z',
-        deleted_at: null,
-        updated_at: '2025-03-05T16:45:26.175Z',
-        identity_id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-        consolidate_number: 0,
-      },
-      view_at: '2025-03-05T16:45:33.770Z',
-      read_at: '2025-03-05T16:45:40.341Z',
-      updated_at: '2025-03-05T16:45:40.341Z',
-      created_at: '2025-03-05T16:45:26.228Z',
-      silent: false,
-    },
-    {
-      id: 'bd5c75ef-9dc0-4516-823c-1dda043060b2',
-      type: 'CHAT',
-      ref_id: '7894ed98-c9d5-4a63-84b8-f26681b77db4',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        id: '7894ed98-c9d5-4a63-84b8-f26681b77db4',
-        body: {
-          body: 'OrganizationTest sent you a new message',
-          title: 'New Message',
-        },
-        text: 'hello shant',
-        type: 'CHAT',
-        media: null,
-        muted: false,
-        orgin: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          created_at: '2025-02-27T10:59:24.278Z',
-        },
-        refId: '7894ed98-c9d5-4a63-84b8-f26681b77db4',
-        chat_id: '47909b77-87b4-461d-b242-eb63bd498fab',
-        replied: false,
-        identity: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T13:16:43.884Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: '47909b77-87b4-461d-b242-eb63bd498fab',
-        reply_id: null,
-        created_at: '2025-03-05T16:44:08.859Z',
-        deleted_at: null,
-        updated_at: '2025-03-05T16:44:08.859Z',
-        identity_id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-        consolidate_number: 0,
-      },
-      view_at: '2025-03-05T16:44:44.101Z',
-      read_at: '2025-03-05T16:44:46.662Z',
-      updated_at: '2025-03-05T16:44:46.662Z',
-      created_at: '2025-03-05T16:44:08.912Z',
-      silent: false,
-    },
-    {
-      id: 'bb12271f-7f3a-4e29-9109-2216f9bcdab7',
-      type: 'CHAT',
-      ref_id: 'bc3957da-2306-436f-bfde-00d3977ce668',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        id: 'bc3957da-2306-436f-bfde-00d3977ce668',
-        body: {
-          body: 'OrganizationTest sent you a new message',
-          title: 'New Message',
-        },
-        text: 'hi',
-        type: 'CHAT',
-        media: null,
-        muted: false,
-        orgin: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          created_at: '2025-02-27T10:59:24.278Z',
-        },
-        refId: 'bc3957da-2306-436f-bfde-00d3977ce668',
-        chat_id: '47909b77-87b4-461d-b242-eb63bd498fab',
-        replied: false,
-        identity: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T13:16:43.884Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: '47909b77-87b4-461d-b242-eb63bd498fab',
-        reply_id: null,
-        created_at: '2025-03-05T16:43:21.085Z',
-        deleted_at: null,
-        updated_at: '2025-03-05T16:43:21.085Z',
-        identity_id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-        consolidate_number: 0,
-      },
-      view_at: '2025-03-05T16:43:35.235Z',
-      read_at: '2025-03-05T16:43:49.440Z',
-      updated_at: '2025-03-05T16:43:49.440Z',
-      created_at: '2025-03-05T16:43:21.141Z',
-      silent: false,
-    },
-    {
-      id: '42c5f1f4-c775-44fa-b2d3-eff308091d7f',
-      type: 'CHAT',
-      ref_id: '374fa3f9-1d1b-40d0-9d06-68f814df1a1b',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        id: '374fa3f9-1d1b-40d0-9d06-68f814df1a1b',
-        body: {
-          body: 'imshantik@gmail.com sent you a new message',
-          title: 'New Message',
-        },
-        text: 'hello organization',
-        type: 'CHAT',
-        media: null,
-        muted: false,
-        orgin: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          created_at: '2025-02-27T13:16:43.884Z',
-        },
-        refId: '374fa3f9-1d1b-40d0-9d06-68f814df1a1b',
-        chat_id: '47909b77-87b4-461d-b242-eb63bd498fab',
-        replied: false,
-        identity: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T10:59:24.278Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: '47909b77-87b4-461d-b242-eb63bd498fab',
-        reply_id: null,
-        created_at: '2025-03-05T16:42:23.356Z',
-        deleted_at: null,
-        updated_at: '2025-03-05T16:42:23.356Z',
-        identity_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-        consolidate_number: 0,
-      },
-      view_at: '2025-03-05T16:42:40.223Z',
-      read_at: '2025-03-05T16:42:50.706Z',
-      updated_at: '2025-03-05T16:42:50.706Z',
-      created_at: '2025-03-05T16:42:23.421Z',
-      silent: false,
-    },
-    {
-      id: '36344c02-2f83-4060-ac86-4fd8353eb493',
-      type: 'OFFER',
-      ref_id: '592f9af4-22dd-4ae9-b7a1-8d491251e506',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        body: {
-          body: 'OrganizationTest sent you an offer',
-          title: 'Have an offer',
-        },
-        type: 'OFFER',
-        orgin: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          created_at: '2025-02-27T10:59:24.278Z',
-        },
-        refId: '592f9af4-22dd-4ae9-b7a1-8d491251e506',
-        identity: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T13:16:43.884Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: '25e61193-9857-4af0-a146-6e7943894ccd',
-        consolidate_number: 0,
-      },
-      view_at: '2025-03-05T12:30:58.257Z',
-      read_at: '2025-03-05T12:31:06.545Z',
-      updated_at: '2025-03-05T12:31:06.545Z',
-      created_at: '2025-03-05T12:30:49.217Z',
-      silent: false,
-    },
-    {
-      id: 'ea8244c4-0253-47e2-a1ca-1a3f33af66bb',
-      type: 'APPLICATION',
-      ref_id: '8843fce6-042e-46dc-b1ae-50bb7ddd5ced',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        body: {
-          body: 'imshantik@gmail.com applied to your job',
-          title: 'Job applied',
-        },
-        type: 'APPLICATION',
-        orgin: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          created_at: '2025-02-27T13:16:43.884Z',
-        },
-        refId: '8843fce6-042e-46dc-b1ae-50bb7ddd5ced',
-        identity: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T10:59:24.278Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: '25e61193-9857-4af0-a146-6e7943894ccd',
-        consolidate_number: 0,
-      },
-      view_at: '2025-03-05T12:12:49.063Z',
-      read_at: '2025-03-05T12:13:57.978Z',
-      updated_at: '2025-03-05T12:13:57.978Z',
-      created_at: '2025-03-05T12:10:48.703Z',
-      silent: false,
-    },
-    {
-      id: '5a45fbc4-04d7-4141-857f-c9ec4f83aa58',
-      type: 'OFFER',
-      ref_id: '2da2348c-f5ae-4440-b9df-cd4629f0ecc6',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        body: {
-          body: 'OrganizationTest sent you an offer',
-          title: 'Have an offer',
-        },
-        type: 'OFFER',
-        orgin: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          created_at: '2025-02-27T10:59:24.278Z',
-        },
-        refId: '2da2348c-f5ae-4440-b9df-cd4629f0ecc6',
-        identity: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T13:16:43.884Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: 'ecce2dec-35a2-4a6a-a0d8-155cd898f6b9',
-        consolidate_number: 0,
-      },
-      view_at: '2025-03-04T09:36:55.895Z',
-      read_at: '2025-03-04T09:37:08.918Z',
-      updated_at: '2025-03-04T09:37:08.918Z',
-      created_at: '2025-03-04T09:36:26.354Z',
-      silent: false,
-    },
-    {
-      id: '0655a77e-e65d-4641-8ecf-5460aa651266',
-      type: 'CHAT',
-      ref_id: '521a71af-d4f3-4434-9fd4-095965bff6c1',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        id: '521a71af-d4f3-4434-9fd4-095965bff6c1',
-        body: {
-          body: 'OrganizationTest sent you a new message',
-          title: 'New Message',
-        },
-        text: 'things are nice here',
-        type: 'CHAT',
-        media: null,
-        muted: false,
-        orgin: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          created_at: '2025-02-27T10:59:24.278Z',
-        },
-        refId: '521a71af-d4f3-4434-9fd4-095965bff6c1',
-        chat_id: '47909b77-87b4-461d-b242-eb63bd498fab',
-        replied: false,
-        identity: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T13:16:43.884Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: '47909b77-87b4-461d-b242-eb63bd498fab',
-        reply_id: null,
-        created_at: '2025-02-28T10:45:03.832Z',
-        deleted_at: null,
-        updated_at: '2025-02-28T10:45:03.832Z',
-        identity_id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-        consolidate_number: 0,
-      },
-      view_at: '2025-02-28T10:45:14.280Z',
-      read_at: '2025-02-28T10:58:47.888Z',
-      updated_at: '2025-02-28T10:58:47.888Z',
-      created_at: '2025-02-28T10:45:04.137Z',
-      silent: false,
-    },
-    {
-      id: 'b7363a62-b77f-435d-8c1d-5499ede4830a',
-      type: 'CHAT',
-      ref_id: '1c59246b-8632-45db-ae9a-5e575cf9574d',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        id: '1c59246b-8632-45db-ae9a-5e575cf9574d',
-        body: {
-          body: 'imshantik@gmail.com sent you a new message',
-          title: 'New Message',
-        },
-        text: 'whats good?',
-        type: 'CHAT',
-        media: null,
-        muted: false,
-        orgin: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          created_at: '2025-02-27T13:16:43.884Z',
-        },
-        refId: '1c59246b-8632-45db-ae9a-5e575cf9574d',
-        chat_id: '47909b77-87b4-461d-b242-eb63bd498fab',
-        replied: false,
-        identity: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T10:59:24.278Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: '47909b77-87b4-461d-b242-eb63bd498fab',
-        reply_id: null,
-        created_at: '2025-02-28T10:38:39.353Z',
-        deleted_at: null,
-        updated_at: '2025-02-28T10:38:39.353Z',
-        identity_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-        consolidate_number: 0,
-      },
-      view_at: '2025-02-28T10:38:47.753Z',
-      read_at: '2025-02-28T10:58:47.888Z',
-      updated_at: '2025-02-28T10:58:47.888Z',
-      created_at: '2025-02-28T10:38:39.404Z',
-      silent: false,
-    },
-    {
-      id: 'ec72432e-098c-43ae-85a7-fbe31cb35bcf',
-      type: 'CHAT',
-      ref_id: '9fb4d3f5-1e6a-44f6-a1b6-67fc6506c7f1',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        id: '9fb4d3f5-1e6a-44f6-a1b6-67fc6506c7f1',
-        body: {
-          body: 'imshantik@gmail.com sent you a new message',
-          title: 'New Message',
-        },
-        text: 'hi',
-        type: 'CHAT',
-        media: null,
-        muted: false,
-        orgin: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          created_at: '2025-02-27T13:16:43.884Z',
-        },
-        refId: '9fb4d3f5-1e6a-44f6-a1b6-67fc6506c7f1',
-        chat_id: '47909b77-87b4-461d-b242-eb63bd498fab',
-        replied: false,
-        identity: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T10:59:24.278Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: '47909b77-87b4-461d-b242-eb63bd498fab',
-        reply_id: null,
-        created_at: '2025-02-28T10:36:55.040Z',
-        deleted_at: null,
-        updated_at: '2025-02-28T10:36:55.040Z',
-        identity_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-        consolidate_number: 0,
-      },
-      view_at: '2025-02-28T10:37:06.632Z',
-      read_at: '2025-02-28T10:58:47.888Z',
-      updated_at: '2025-02-28T10:58:47.888Z',
-      created_at: '2025-02-28T10:36:55.090Z',
-      silent: false,
-    },
-    {
-      id: '07894516-ba0c-4aab-a6eb-e6184bdaab1f',
-      type: 'CHAT',
-      ref_id: '09a33391-d6fe-47f1-aeff-f5fb307a0605',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        id: '09a33391-d6fe-47f1-aeff-f5fb307a0605',
-        body: {
-          body: 'OrganizationTest sent you a new message',
-          title: 'New Message',
-        },
-        text: 'hello',
-        type: 'CHAT',
-        media: null,
-        muted: false,
-        orgin: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          created_at: '2025-02-27T10:59:24.278Z',
-        },
-        refId: '09a33391-d6fe-47f1-aeff-f5fb307a0605',
-        chat_id: '47909b77-87b4-461d-b242-eb63bd498fab',
-        replied: false,
-        identity: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T13:16:43.884Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: '47909b77-87b4-461d-b242-eb63bd498fab',
-        reply_id: null,
-        created_at: '2025-02-28T10:36:25.925Z',
-        deleted_at: null,
-        updated_at: '2025-02-28T10:36:25.925Z',
-        identity_id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-        consolidate_number: 0,
-      },
-      view_at: '2025-02-28T10:36:32.705Z',
-      read_at: '2025-02-28T10:58:47.888Z',
-      updated_at: '2025-02-28T10:58:47.888Z',
-      created_at: '2025-02-28T10:36:25.987Z',
-      silent: false,
-    },
-    {
-      id: 'e3a5cfa7-2504-43fd-b5e4-869e6932edef',
-      type: 'APPLICATION',
-      ref_id: 'cfb626d1-2bf5-474a-bd78-3ddf41b81669',
-      user_id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      data: {
-        body: {
-          body: 'imshantik@gmail.com applied to your job',
-          title: 'Job applied',
-        },
-        type: 'APPLICATION',
-        orgin: {
-          id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-          meta: {
-            id: '1396c403-1ade-4c40-b0d1-4680059a46c5',
-            city: 'Tehran',
-            name: 'OrganizationTest',
-            email: 'imshantik@gmal.com',
-            image: null,
-            hiring: false,
-            status: 'ACTIVE',
-            address: null,
-            country: 'IR',
-            mission: null,
-            verified: false,
-            shortname: 'organizationtest',
-            description: null,
-            wallet_address: null,
-            verified_impact: false,
-          },
-          type: 'organizations',
-          created_at: '2025-02-27T13:16:43.884Z',
-        },
-        refId: 'cfb626d1-2bf5-474a-bd78-3ddf41b81669',
-        identity: {
-          id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-          meta: {
-            id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-            city: null,
-            name: null,
-            email: 'imshantik@gmail.com',
-            avatar: null,
-            status: 'ACTIVE',
-            address: null,
-            country: null,
-            username: 'imshantik5140',
-            open_to_work: true,
-            is_contributor: null,
-            wallet_address: null,
-            identity_verified: false,
-            open_to_volunteer: false,
-          },
-          type: 'users',
-          follower: false,
-          following: false,
-          created_at: '2025-02-27T10:59:24.278Z',
-          connection_status: null,
-          verification_status: null,
-        },
-        parentId: 'ecce2dec-35a2-4a6a-a0d8-155cd898f6b9',
-        consolidate_number: 0,
-      },
-      view_at: '2025-02-28T10:24:00.022Z',
-      read_at: '2025-02-28T10:58:47.888Z',
-      updated_at: '2025-02-28T10:58:47.888Z',
-      created_at: '2025-02-28T10:23:52.952Z',
-      silent: false,
-    },
-  ],
-};
