--- cypress/e2e/communities/postMock.ts
@@ -1,90 +1,0 @@
-export const POST = {
-  id: 'bf59fb3f-c07c-465d-9303-33a15ab507ae',
-  content: 'content',
-  identity_id: '20b07c1c-b142-4ee3-a87f-122aa99a7659',
-  created_at: '2024-11-18T19:22:56.439Z',
-  updated_at: '2024-11-18T19:22:56.439Z',
-  deleted_at: null,
-  causes_tags: ['POVERTY'],
-  hashtags: null,
-  identity_tags: null,
-  media: [
-    {
-      id: '0ae176a0-22f4-4e89-9d2a-5f89b6cbe191',
-      url: 'https://storage.googleapis.com/socious-gcs//060ee16295da3a568d5b2738754e16fe.png',
-    },
-  ],
-  likes: 2,
-  shared: 0,
-  shared_id: null,
-  old_id: null,
-  search_tsv: "'content':1 'poverti':2",
-  title: 'test2',
-  comments: 0,
-  identity_type: 'users',
-  identity_meta: {
-    id: '20b07c1c-b142-4ee3-a87f-122aa99a7659',
-    city: 'Tehran, Tehran Province',
-    name: 'Sanaz Mahmoudi',
-    email: 'mahmoudisanaz59@gmail.com',
-    avatar: 'https://storage.googleapis.com/socious-gcs/40feef87c121271bfb9b5b26d9748219.jpg',
-    status: 'ACTIVE',
-    address: null,
-    country: 'IR',
-    username: 'mahmoudisanaz599738',
-    open_to_work: false,
-    is_contributor: null,
-    wallet_address: '0x6Ff5d5EC0eb3EE0FfeE5C9238bC1266afB56Ec4d',
-    identity_verified: false,
-    open_to_volunteer: false,
-  },
-  liked: true,
-  reported: false,
-  shared_post: null,
-  shared_from_identity: null,
-  emojis: null,
-  liked_identities: [
-    {
-      id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-      meta: {
-        id: '035f724e-5824-4129-8ed6-96c6dda81a3a',
-        city: null,
-        name: null,
-        email: 'imshantik@gmail.com',
-        avatar: null,
-        status: 'ACTIVE',
-        address: null,
-        country: null,
-        username: 'imshantik5140',
-        open_to_work: true,
-        is_contributor: null,
-        wallet_address: null,
-        identity_verified: false,
-        open_to_volunteer: false,
-      },
-      type: 'users',
-      created_at: '2025-02-27T10:59:24.278886+00:00',
-    },
-    {
-      id: '48ef808d-85eb-4e6d-a8a5-853cd93abc4d',
-      meta: {
-        id: '48ef808d-85eb-4e6d-a8a5-853cd93abc4d',
-        city: 'Arinsal',
-        name: 'Nina N',
-        email: 'testazintest2@gmail.com',
-        avatar: null,
-        status: 'ACTIVE',
-        address: null,
-        country: 'AD',
-        username: 'testazintest2',
-        open_to_work: false,
-        is_contributor: true,
-        wallet_address: '0x6990bf996E984cffde6EbCAb6AC8723c95069719',
-        identity_verified: false,
-        open_to_volunteer: false,
-      },
-      type: 'users',
-      created_at: '2022-11-03T06:22:49.394579+00:00',
-    },
-  ],
-};
